Issue#11: Simplify configuration
This change removes all the unchanged configuration items in
`config.local.php`, so it's easier to spot what the differences are to
the default config.
Then, at the end, it checks for a `config.site.php` and includes it if
it exist. This way configuration changes can be made there instead of
the default files, making changes easier to manage.

Issue #11